Big architecture frontend refactoring Part 1. (radical core / framework separation).
Core will include most of the code and all of the business logic. Framework will only do simple, dumm rendering, and call core's methods when user clicks or something. NOTHING MORE.
This radical separation of conerns will make us flexible and independent of any framework. Our frontend app is just an app. We can take it, and make a CLI nodejs program if we wish to. Core will stay identical: it will not change in the slightest. We will just have to write a new UI part, which, as mentioned, is small, simple and has no business logic whatsoever.